% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totals.R
\name{totals}
\alias{totals}
\title{Total Time Visualizations in R}
\usage{
totals(df, act)
totals(df, act, colors)
}
\arguments{
\item{df}{json data with keys as numbers (as characters) and values for percent of total time.}

\item{colors}{3 x n (number of activities) json array of colors for the graph. Each hex color in sub-array represents:
\enumerate{
\item color for the wave and border, 2) color for text when under the wave, 3) color for the text.
}}

\item{act}{json array of the names of activities.}
}
\description{
Draws a visualizations of total time spent on different activities in the sample as percent of total time.
}
\references{
Kolpashnikova, Kamila. (2022). Total Time Visualizations in R. Toronto,ON: York University.
}
